Code is ready for submission. mfs.c renamed to msh.c. Detailed comments added.

+++ proc.c
@@ -1,3 +1,13 @@
+/*
+ * Name: Shaun Sartin
+ * ID #: 1000992189
+ * Programming Assignment 1 (Proc)
+ * Description: This program has 2 modes, default-mode and verbose-mode.
+ * Default-mode will display some system info and then terminate.
+ * Verbose-mode will display more system info every specified interval and
+ * will not terminate unless forced to do so. (For Example: Ctrl+C)
+ */
+
 #include <stdio.h>
 #include <stdlib.h>
 #include <unistd.h>
@@ -5,11 +15,34 @@
 
 #define MAX_STRING_LENGTH 255
 
-//TODO ADD Description
-void printProc(float userTime, float systemTime, float idleTime, int ctxtSwitches,
-  int processCreations, float totalMem, float freeMem, float diskStats, float uptime)
+/*
+ * Function: printProc
+ * Parameter 1: userTime - A float representing the time (in milliseconds) that
+ * the system has spent in user mode.
+ * Parameter 2: systemTime - A float representing the time (in milliseconds) that
+ * the system has spent in system mode.
+ * Parameter 3: systemTime - A float representing the time (in milliseconds) that
+ * the system has spent idling.
+ * Parameter 4: ctxtSwitches - A float representing the number of context switches
+ * that have taken place since the system's last boot.
+ * Parameter 5: processCreations - A float representing the number of processes
+ * that have been created since the system's last boot.
+ * Parameter 6: totalMem - A float representing the total amount of physical RAM.
+ * Parameter 7: freeMem - A float representing the amount of physical RAM unused
+ * by the system.
+ * Parameter 8: diskStats - A float representing the number of disk reads/writes
+ * since the system's last boot.
+ * Parameter 9: uptime - A float representing the number of seconds since the system's
+ * last boot.
+ * Description: This function is only used in the verbose mode.
+ * It calculates the percent values and average rates since system uptime
+ * It also prints all of the calculated values, except for totalTime, to stdout.
+ */
+void printProc(float userTime, float systemTime, float idleTime, float ctxtSwitches,
+  float processCreations, float totalMem, float freeMem, float diskStats, float uptime)
 {
   float totalTime = userTime + systemTime + idleTime;
+
   float userPercent = (userTime / totalTime) * 100;
   float systemPercent = (systemTime / totalTime) * 100;
   float idlePercent = (idleTime / totalTime) * 100;
@@ -31,24 +64,48 @@ void printProc(float userTime, float systemTime, float idleTime, int ctxtSwitche
 int main(int argc, char** argv)
 {
 
+  // This is used to open all the necessary files in /proc
   FILE * filePointer;
+
+  // These strings hold information that is displayed in 'default-mode'
   char processorString[MAX_STRING_LENGTH];
   char kernelString[MAX_STRING_LENGTH];
   char totalMemString[MAX_STRING_LENGTH];
   char uptimeString[MAX_STRING_LENGTH];
+
+  // This string is overwritten many times. It is used as a place to dump
+  // irrelevant information, as well as perform string comparisons for
+  // certain strings in the /proc files.
   char dummyString[MAX_STRING_LENGTH];
 
+  // This variable will hold the time (in seconds) since the system has last booted
   int uptime;
+
+  // These variables contain information from /proc/stat.
+  // They are only used in 'verbose-mode'
   int userTime;
   int systemTime;
   int idleTime;
   int ctxtSwitches;
   int processCreations;
+
+  // These variables hold information from /proc/meminfo
+  // totalMem is used in both modes, whereas freeMem is only used in 'verbose-mode'
   int totalMem;
   int freeMem;
+
+  // This variable contains the number of disk sectors that have been read
+  // That information is pulled from /proc/diskstats
   int diskStats;
+
+  // This variable is set via an argument when executing this file
+  // Therefore, it is only relevant for 'verbose-mode'.
+  // It determines how long to wait before performing calculations and printing.
   int sleepTime;
 
+
+  // If an argument has been passed, run 'verbose-mode'
+  // 'verbose-mode' refers to everything in this if-statement.
   if (argv[1] != 0)
   {
     sleepTime = strtol(argv[1], NULL, 10);
@@ -162,7 +219,7 @@ int main(int argc, char** argv)
     }
   }
 
-  // This 'else' block holds the default functionality of the program
+  // This 'else' block holds the 'default-mode' functionality of the program
   else
   {
     //Start working with data in cpuinfo file
